Introduced a health check.
When PINGed trhough its frontend socket, the broker now sends back
its PID.

This is used to verify that there's no running (and working) broker
before starting a new one.

This is also a health check that can be used by Nagios to
make sure the broker is up and running:

    >>> from powerhose.util import verify_broker
    >>> verify_broker('ipc://here')
    '1213213'

If there's no running broker, returns None.